% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wahlergebnisse-data.R
\docType{data}
\name{wahlergebnisse}
\alias{wahlergebnisse}
\title{Wahlergebnisse der Bundestagswahl 1990 bis 2017.}
\format{
An object of class \code{"data.frame"}; see \code{\link[base]{data.frame}}.
}
\usage{
data(wahlergebnisse)
}
\description{
Ausgewählte Parteien in Prozent.
}
\examples{
data(wahlergebnisse)
}
\keyword{datasets}
